reorg nodedb
